New namespacing. Abstract type gem. Lazy detail parsing.
