+++ Gemfile
@@ -4,6 +4,7 @@ gem 'nokogiri'
 gem 'curb'
 gem 'dotenv'
 gem 'discourse_api'
+gem 'abstract_type'
 
 group :development do
   gem 'rspec'
