--- lib/notifiers.rb
@@ -1,6 +1,0 @@
-require_relative 'notifiers/base'
-require_relative 'notifiers/slack'
-require_relative 'notifiers/discourse'
-
-module Notifiers
-end
