+++ lib/uvobot/discourse_client.rb
@@ -1,0 +1,12 @@
+require 'discourse_api'
+
+module Uvobot
+  class DiscourseClient < DiscourseApi::Client
+    def create_topic(args = {})
+      super(args)
+    rescue DiscourseApi::Error => e
+      # puts e.message
+      return nil
+    end
+  end
+end
