+++ lib/uvobot/notifications.rb
@@ -1,0 +1,8 @@
+require_relative 'notifications/base'
+require_relative 'notifications/slack_notifier'
+require_relative 'notifications/discourse_notifier'
+
+module Uvobot
+  module Notifications
+  end
+end
